fix: set default `importJavaTime` value to `false`
The `importJavaTime` setting modifies the import behavior for Java Date/Time classes by applying a config script to the Groovy compiler. This causes issues in Grails plugins, as the config script interferes with the generation of the `META-INF/grails-plugin.xml` file.

By defaulting `importJavaTime` to `false`, these issues are avoided until a more robust solution can be implemented.