Fix `grails-plugin.xml` generation and `importJavaTime` setting (#388)
* fix: set default `importJavaTime` value to `false`

The `importJavaTime` setting modifies the import behavior for Java Date/Time classes by applying a config script to the Groovy compiler. This causes issues in Grails plugins, as the config script interferes with the generation of the `META-INF/grails-plugin.xml` file.

By defaulting `importJavaTime` to `false`, these issues are avoided until a more robust solution can be implemented.

* fix: honor the `importJavaTime` setting

This fix addresses a problem where the `importJavaTime` setting was not honored as it was used in the evaluation phase, before the project got a chance to change it.

* build: add the `grails-bom`

The `grails-bom` is no longer an `api` dependency in `grails-core`.

* fix: support multiple compiler configs to be merged

This commit will enable more than one compiler configuration to be used by merging them. It will also merge the compiler configs for generating `grails-plugin.xml` and importing Java Date/Time.